requirements.txt
autogluon.tabular==1.5.0
optuna>=3.5.0

pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

lightgbm>=4.0.0
xgboost>=2.0.0
catboost>=1.2.0

matplotlib>=3.7.0
seaborn>=0.12.0

jupyter